audio-io fixed build issue on 64-bit platform
[Version] 0.3.6
[Profile] Common
[Issue Type] Bug
[Dependency module] NA
[Dependency commit] NA
[Comment]
'size_t' uses 8 bytes in 64-bit platform.

+++ include/CAudioInput.h
@@ -76,7 +76,7 @@ namespace tizen_media_audio {
         virtual void setStreamCallback(SStreamCallback callback) throw (CAudioError);
 
         /* Methods */
-        int read(void* buffer, size_t length) throw (CAudioError);
+        size_t read(void* buffer, size_t length) throw (CAudioError);
         int peek(const void** buffer, size_t* length) throw (CAudioError);
         int drop() throw (CAudioError);
     };
